# specs/openapi/ai-service.yaml
openapi: 3.1.0
info:
  title: Poetry RAG — AI Service API
  version: "0.4.0"
  description: |
    FastAPI AI 微服务对外暴露的接口。
    内部集成 LangGraph ReAct Agent、可配置 LLM（GLM / 豆包 / OpenAI / Ollama 等）、ChromaDB 向量检索、图像生成（CogView-4 等）。
    提供健康检查、异步生成、同步冒烟测试、思考流 SSE、LangGraph 对话 SSE、分镜多图 SSE 六类接口。
    LLM 和图像模型通过环境变量 `LLM_PROVIDER` / `IMAGE_PROVIDER` 自由配置，无需修改代码。
servers:
  - url: http://localhost:8000
    description: 本地开发
...
paths:
...
  /ai/api/v1/generate/async:
    post:
      summary: 异步启动 RAG 生成管道
      operationId: queueGeneration
      tags: [generate]
      description: |
        接收任务信息后立即返回 202，
        在后台执行 RAG 检索 → Prompt 增强 → 图像生成，
        完成后以 `POST callbackUrl` 回调 Backend。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerationRequest'
            example:
              taskId: "abc123-uuid"
              sourceText: "床前明月光，疑是地上霜"
              callbackUrl: "http://backend:8080/api/v1/poetry/callback"
              callbackToken: "poetry-callback-token-change-me"
      responses:
        "202":
          description: 任务已接受，后台处理中
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task accepted"
                  taskId:
                    type: string

  # ── 同步简单生成（冒烟测试）───────────────────────────────────────────────
  /ai/api/v1/generate/simple:
    post:
      summary: 同步生成（本地冒烟用，无回调）
      operationId: simpleGeneration
      tags: [generate]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleGenerationRequest'
      responses:
        "200":
          description: 生成结果（同步返回）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleGenerationResponse'

  # ── GLM 思考流 SSE ────────────────────────────────────────────────────────
  /ai/api/v1/generate/think-stream:
    post:
      summary: GLM 推理过程 SSE 流
      operationId: thinkStream
      tags: [generate]
      description: |
        逐 token 推送 GLM 的 `reasoning_content`。
        适合前端实时展示"思考过程"。
        每个 SSE 事件格式：`data: {"text":"..."}`
        结束标记：`data: [DONE]`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleGenerationRequest'
      responses:
        "200":
          description: Server-Sent Events 流
          content:
            text/event-stream:
              schema:
                type: string

  # ── 对话 Session 管理 ──────────────────────────────────────────────────────
...
  /ai/api/v1/chat/session/{session_id}/history:
    get:
      summary: 获取会话历史消息
      operationId: getChatHistory
      tags: [chat]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 历史消息列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatHistoryItem'
        "503":
          description: LLM_API_KEY 未配置（或 LLM_PROVIDER 对应密钥缺失）
  /ai/api/v1/chat/session/{session_id}:
    delete:
      summary: 清空会话历史记忆
      operationId: clearChatSession
      tags: [chat]
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 会话已清空
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  session_id:
                    type: string

  # ── LangGraph Agent 对话 SSE ──────────────────────────────────────────────
  /ai/api/v1/chat:
    post:
      summary: LangGraph ReAct Agent 流式对话
      operationId: chatStream
      tags: [chat]
      description: |
        基于 LangGraph ReAct Agent（可配置 LLM + ChromaDB）的 SSE 流式对话接口。

        Agent 拥有两个工具：
        - `search_poetry`：在 ChromaDB 向量库中检索古诗词
        - `visualize_poem`：RAG → Prompt 增强 → 图像生成（取决于 IMAGE_PROVIDER 配置）

        **SSE 事件格式**（每行 `data: {JSON}`）：
        | type | 触发时机 | 关键字段 |
        |------|---------|----------|
        | `token` | LLM 逐 token 输出 | `content` |
        | `tool` | 工具调用开始 | `name`, `display`, `input` |
        | `tool_end` | 工具调用返回 | `name`, `output` |
        | `node_progress` | Agent/Tools 节点切换 | `node`, `iteration`, `label` |
        | `rag_result` | search_poetry 工具完成时额外推送 | `mode`, `poems` |
        | `done` | 流结束 | `session_id` |
        | `error` | 发生错误 | `content` |
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              message: "李白最著名的诗是哪首？"
              session_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Server-Sent Events 流
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE 事件流，可能的事件示例：
                  data: {"type": "token", "content": "李"}
                  data: {"type": "tool", "name": "search_poetry", "display": "正在检索古诗词库……", "input": "{\"keyword\": \"李白\"}"}
                  data: {"type": "tool_end", "name": "search_poetry", "output": "**《静夜思》…"}
                  data: {"type": "rag_result", "mode": "exact", "poems": [{"title": "静夜思", ...}]}
                  data: {"type": "done", "session_id": "550e8400..."}
        "503":
          description: LLM_API_KEY 未配置（或 LLM_PROVIDER 对应密钥缺失），Agent 不可用

  # ── 分镜多图生成 SSE ──────────────────────────────────────────────────────
  /ai/api/v1/generate/storyboard:
    post:
      summary: 诗词分镜多图生成（SSE）
      operationId: storyboardStream
      tags: [storyboard]
      description: |
        输入一句诗词，系统自动执行：
        1. **RAG 检索**：从 ChromaDB 向量库获取完整原诗 + 白话译文
        2. **LLM 规划**：生成 1 张全景基调图 + 3-5 张叙事分镜方案（取决于 LLM_PROVIDER 配置）
        3. **图像生成逐张推送**：每完成一张立即推送（图与图之间 4s 限速间隔，取决于 IMAGE_PROVIDER 配置）

        **SSE 事件格式**（每行 `data: {JSON}`）：
        | type | 触发时机 | 关键字段 |
        |------|---------|----------|
        | `progress` | RAG 完成 / 开始规划 / 等待限速 | `stage`, `message`, `current`, `total` |
        | `plan` | GLM 规划方案生成完毕 | `poem_title`, `global_style`, `total_shots`, `shots` |
        | `shot_done` | 单张图生成完成 | `shot_id`, `shot_name`, `shot_type`, `poem_lines`, `image_url`, `current`, `total` |
        | `shot_error` | 单张图生成失败 | `shot_id`, `shot_name`, `message`, `current`, `total` |
        | `done` | 全部图生成完毕 | `total_shots` |

        `progress.stage` 枚举值：`rag_done` / `planning` / `generating` / `waiting`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleGenerationRequest'
            example:
              sourceText: "大漠孤烟直，长河落日圆"
      responses:
        "200":
          description: Server-Sent Events 流
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE 事件流示例：
                  data: {"type": "progress", "stage": "rag_done", "message": "已检索到《使至塞上》…", "current": 0, "total": 0}
                  data: {"type": "plan", "poem_title": "使至塞上", "global_style": "大漠苍凉…", "total_shots": 4, "shots": [...]}
                  data: {"type": "shot_done", "shot_id": 1, "shot_name": "全景", "image_url": "http://…", "current": 1, "total": 4}
                  data: {"type": "done", "total_shots": 4}
        "503":
          description: LLM_API_KEY 未配置（或 LLM_PROVIDER 对应密钥缺失），分镜服务不可用
        "500":
          description: RAG 检索失败

  # ── Mock 调试接口 ─────────────────────────────────────────────────────────
...
components:
  schemas:
    GenerationRequest:
      type: object
      required: [taskId, sourceText, callbackUrl, callbackToken]
      properties:
        taskId:
          type: string
          description: Backend 分配的任务 ID
        sourceText:
          type: string
          description: 原始诗句文本
        callbackUrl:
          type: string
          format: uri
          description: 完成后 POST 回调的地址
        callbackToken:
          type: string
          description: 回调鉴权 token，放入 X-Callback-Token 头
    SimpleGenerationRequest:
      type: object
      required: [sourceText]
      properties:
        sourceText:
          type: string
          description: 原始诗句文本
    SimpleGenerationResponse:
      type: object
      properties:
        retrievedText:
          type: string
          description: RAG 检索到的相关语料
        enhancedPrompt:
          type: string
          description: 增强后的图像生成 Prompt
        imageUrl:
          type: string
          format: uri
          description: 生成图像的访问 URL
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: 用户输入的问题或指令
          example: "李白最著名的诗是哪首？"
        session_id:
          type: string
          description: 会话 ID，留空时自动创建新会话（done 事件中返回实际 session_id）
          default: ""
...
    ChatHistoryItem:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string

    # ── 分镜相关 Schema ───────────────────────────────────────────────────
...